Issue 69.2: extract daemon lifecycle to server/lifecycle.rs
Final step of Issue 69. main.rs goes from 317 → 55 prod LOC by
relocating the daemon lifecycle (run + parse_args + companion
socket-path helpers) into a new sibling module
userspace-dp/src/server/lifecycle.rs.

What moved:
- pub(crate) fn run (194 LOC) — the daemon driver: parses args,
  builds the initial Coordinator + ServerState, opens the control
  socket, runs the accept-and-dispatch loop forever
- pub(crate) fn parse_args (48 LOC)
- pub(crate) fn derive_session_socket_path (8 LOC)
- pub(crate) fn derive_event_socket_path (8 LOC)

All widened from file-private `fn` to `pub(crate) fn` so main.rs's
shell can call `server::lifecycle::run()`.

server/mod.rs declares the new module:
    pub(crate) mod lifecycle;

main.rs is now a strict entry-point shell — module declarations,
use statements, and a 5-line `fn main()` that dispatches to
`server::lifecycle::run()`:

    fn main() {
        if let Err(err) = server::lifecycle::run() {
            eprintln!("xpf-userspace-dp: {err}");
            std::process::exit(1);
        }
    }

Pure relocation. Bodies byte-for-byte identical.

cargo test --release: 865 passed, 0 failed (warning count 92 baseline).

Cumulative Issue 69 result:
- main.rs: 845 → 55 prod LOC
- server/helpers.rs: 540 (Issue 69.1, #1073)
- server/lifecycle.rs: 278 (Issue 69.2, this PR)
