+++ userspace-dp/src/server/lifecycle.rs
@@ -1,0 +1,278 @@
+// Daemon lifecycle: argument parsing, control-socket setup, daemon
+// supervision loop. Extracted from main.rs (Issue 69.2).
+//
+// `run` is the main daemon driver — initializes Args, builds the
+// initial Coordinator + ServerState, opens the control socket, and
+// runs the accept-and-dispatch loop forever.
+// `parse_args` parses the command-line into Args.
+// `derive_session_socket_path` / `derive_event_socket_path` are
+// small helpers that derive companion socket paths from the
+// primary control-socket path.
+//
+// Pure relocation. Bodies byte-for-byte identical; visibility
+// widened from file-private to `pub(crate)` so main.rs's `fn main()`
+// shell can call into `server::lifecycle::run`.
+
+use super::super::*;
+
+pub(crate) fn run() -> Result<(), String> {
+    // Increase socket receive buffer defaults — needed for AF_XDP copy mode
+    // to avoid drops when the kernel backlog is large.
+    for sysctl in &[
+        "/proc/sys/net/core/rmem_default",
+        "/proc/sys/net/core/rmem_max",
+    ] {
+        if let Err(e) = fs::write(sysctl, "16777216") {
+            eprintln!("warn: set {sysctl}: {e}");
+        } else {
+            eprintln!("set {sysctl}=16777216");
+        }
+    }
+    let args = parse_args()?;
+    // Enable NAPI busy polling sysctls only in busy-poll mode.
+    // In interrupt mode, skip these so the kernel uses normal interrupt delivery.
+    if args.poll_mode == PollMode::BusyPoll {
+        for (path, val) in &[
+            ("/proc/sys/net/core/busy_poll", "50"),
+            ("/proc/sys/net/core/busy_read", "50"),
+        ] {
+            if let Err(e) = fs::write(path, val) {
+                eprintln!("warn: set {path}: {e}");
+            } else {
+                eprintln!("set {path}={val}");
+            }
+        }
+    } else {
+        eprintln!("xpf-userspace-dp: interrupt mode — skipping busy_poll sysctls");
+    }
+    if let Some(parent) = Path::new(&args.control_socket).parent() {
+        fs::create_dir_all(parent).map_err(|e| format!("create control dir: {e}"))?;
+    }
+    if let Some(parent) = Path::new(&args.state_file).parent() {
+        fs::create_dir_all(parent).map_err(|e| format!("create state dir: {e}"))?;
+    }
+    let _ = fs::remove_file(&args.control_socket);
+    let session_socket = derive_session_socket_path(&args.control_socket);
+    let _ = fs::remove_file(&session_socket);
+
+    let listener = UnixListener::bind(&args.control_socket)
+        .map_err(|e| format!("listen {}: {e}", args.control_socket))?;
+    listener
+        .set_nonblocking(true)
+        .map_err(|e| format!("set nonblocking listener: {e}"))?;
+
+    let session_listener = UnixListener::bind(&session_socket)
+        .map_err(|e| format!("listen session {}: {e}", session_socket))?;
+    session_listener
+        .set_nonblocking(true)
+        .map_err(|e| format!("set nonblocking session listener: {e}"))?;
+    eprintln!("xpf-userspace-dp: session socket at {}", session_socket);
+
+    let state_writer = Arc::new(StateWriter::new());
+    let running = Arc::new(AtomicBool::new(true));
+    let state = Arc::new(Mutex::new(ServerState {
+        status: ProcessStatus {
+            pid: std::process::id() as i32,
+            started_at: Utc::now(),
+            control_socket: args.control_socket.clone(),
+            state_file: args.state_file.clone(),
+            workers: args.workers,
+            ring_entries: args.ring_entries,
+            helper_mode: "rust-afxdp-bootstrap".to_string(),
+            io_uring_planned: true,
+            io_uring_active: false,
+            io_uring_mode: String::new(),
+            io_uring_last_error: String::new(),
+            enabled: false,
+            forwarding_armed: false,
+            capabilities: UserspaceCapabilities::default(),
+            last_snapshot_generation: 0,
+            last_fib_generation: 0,
+            last_snapshot_at: None,
+            interface_addresses: 0,
+            neighbor_entries: 0,
+            neighbor_generation: 0,
+            route_entries: 0,
+            worker_heartbeats: Vec::new(),
+            worker_runtime: Vec::new(),
+            cos_no_owner_binding_drops_total: 0,
+            per_binding: Vec::new(),
+            ha_groups: Vec::new(),
+            fabrics: Vec::new(),
+            queues: Vec::new(),
+            bindings: Vec::new(),
+            recent_session_deltas: Vec::new(),
+            recent_exceptions: Vec::new(),
+            cos_interfaces: Vec::new(),
+            filter_term_counters: Vec::new(),
+            last_resolution: None,
+            slow_path: SlowPathStatus::default(),
+            debug_worker_threads: 0,
+            debug_identity_slots: 0,
+            debug_live_slots: 0,
+            debug_planned_workers: 0,
+            debug_planned_bindings: 0,
+            debug_reconcile_calls: 0,
+            debug_reconcile_stage: String::new(),
+            event_stream_connected: false,
+            event_stream_seq: 0,
+            event_stream_acked: 0,
+            event_stream_sent: 0,
+            event_stream_dropped: 0,
+            last_cache_flush_at: 0,
+        },
+        snapshot: None,
+        afxdp: {
+            let mut c = afxdp::Coordinator::new();
+            c.poll_mode = args.poll_mode;
+            c
+        },
+        state_writer: state_writer.clone(),
+    }));
+    eprintln!("xpf-userspace-dp: poll_mode={:?}", args.poll_mode);
+
+    // Start the event stream sender (connects to daemon's event listener socket).
+    {
+        let event_socket_path = derive_event_socket_path(&args.control_socket);
+        let mut guard = state.lock().expect("state poisoned");
+        guard.afxdp.start_event_stream(&event_socket_path);
+        eprintln!(
+            "xpf-userspace-dp: event stream targeting {}",
+            event_socket_path
+        );
+    }
+
+    {
+        let running = running.clone();
+        ctrlc::set_handler(move || {
+            running.store(false, Ordering::SeqCst);
+        })
+        .map_err(|e| format!("install ctrlc handler: {e}"))?;
+    }
+
+    write_state(&args.state_file, &state)?;
+
+    // Spawn a dedicated thread for the session socket so session installs
+    // (HA sync path) proceed concurrently with main socket operations
+    // (status polls, snapshot publishes). The shared `state` mutex already
+    // protects concurrent access. Fixes #452.
+    let session_thread = {
+        let state = state.clone();
+        let running = running.clone();
+        let state_file = args.state_file.clone();
+        thread::Builder::new()
+            .name("session-socket".to_string())
+            .spawn(move || {
+                while running.load(Ordering::SeqCst) {
+                    match session_listener.accept() {
+                        Ok((stream, _)) => {
+                            let _ =
+                                handle_stream(stream, &state_file, state.clone(), running.clone());
+                        }
+                        Err(err) if err.kind() == std::io::ErrorKind::WouldBlock => {
+                            thread::sleep(Duration::from_millis(10));
+                        }
+                        Err(err) => {
+                            eprintln!("xpf-userspace-dp: accept session: {err}");
+                            continue;
+                        }
+                    }
+                }
+            })
+            .map_err(|e| format!("spawn session thread: {e}"))?
+    };
+
+    while running.load(Ordering::SeqCst) {
+        match listener.accept() {
+            Ok((stream, _)) => {
+                let _ = handle_stream(stream, &args.state_file, state.clone(), running.clone());
+            }
+            Err(err) if err.kind() == std::io::ErrorKind::WouldBlock => {
+                thread::sleep(Duration::from_millis(50));
+            }
+            Err(err) => return Err(format!("accept: {err}")),
+        }
+    }
+
+    // Wait for the session thread to finish.
+    if let Err(panic) = session_thread.join() {
+        eprintln!("xpf-userspace-dp: session thread panicked: {panic:?}");
+    }
+    {
+        let mut guard = state.lock().expect("state poisoned");
+        guard.afxdp.stop_with_event_stream();
+        refresh_status(&mut guard);
+    }
+    afxdp::remove_kernel_rst_suppression();
+    write_state(&args.state_file, &state)?;
+    let _ = fs::remove_file(&args.control_socket);
+    let _ = fs::remove_file(&session_socket);
+    Ok(())
+}
+
+/// Derive the session socket path from the control socket path.
+/// `/run/xpf/userspace-dp.sock` -> `/run/xpf/userspace-dp-sessions.sock`
+pub(crate) fn derive_session_socket_path(control_socket: &str) -> String {
+    match control_socket.rsplit_once('/') {
+        Some((dir, _)) => format!("{}/userspace-dp-sessions.sock", dir),
+        None => "userspace-dp-sessions.sock".to_string(),
+    }
+}
+
+/// Derive the event socket path from the control socket path.
+/// `/run/xpf/control.sock` -> `/run/xpf/userspace-dp-events.sock`
+pub(crate) fn derive_event_socket_path(control_socket: &str) -> String {
+    match control_socket.rsplit_once('/') {
+        Some((dir, _)) => format!("{dir}/userspace-dp-events.sock"),
+        None => "userspace-dp-events.sock".to_string(),
+    }
+}
+
+pub(crate) fn parse_args() -> Result<Args, String> {
+    let mut control_socket = env::temp_dir()
+        .join("xpf-userspace-dp")
+        .join("control.sock")
+        .to_string_lossy()
+        .to_string();
+    let mut state_file = env::temp_dir()
+        .join("xpf-userspace-dp")
+        .join("state.json")
+        .to_string_lossy()
+        .to_string();
+    let mut workers = 1usize;
+    let mut ring_entries = 4096usize;
+    let mut poll_mode = PollMode::BusyPoll;
+
+    let mut args = env::args().skip(1);
+    while let Some(arg) = args.next() {
+        let val = args
+            .next()
+            .ok_or_else(|| format!("missing value for argument {arg}"))?;
+        match arg.as_str() {
+            "--control-socket" => control_socket = val,
+            "--state-file" => state_file = val,
+            "--workers" => {
+                workers = val
+                    .parse::<usize>()
+                    .map_err(|e| format!("parse --workers: {e}"))?
+                    .max(1)
+            }
+            "--ring-entries" => {
+                ring_entries = val
+                    .parse::<usize>()
+                    .map_err(|e| format!("parse --ring-entries: {e}"))?
+                    .max(1)
+            }
+            "--poll-mode" => poll_mode = PollMode::from_str(&val),
+            other => return Err(format!("unknown argument {other}")),
+        }
+    }
+
+    Ok(Args {
+        control_socket,
+        state_file,
+        workers,
+        ring_entries,
+        poll_mode,
+    })
+}
